calculateMostLikelyAlleles deals with ties correctly
AlleleSubsettingUtils.calculateMostLikelyAlleles was incorrectly removing all alleles if the best alleles were tied and numAltAllelesToKeep=1
fixes #2157

move tests into AlleleSubsettingUtilsUnitTest from GATKVariantContextUtilsUnitTest as part of #2158
a few typo fixes that I noticed
